Use Puppeteer instead of PhantomJS
PhantomJS is out and Puppeteer is in. Puppeteer is a headless Chrome
Node API from Google. It is maintained and is much quicker. We have
therefore removed the need for pooling since Puppeteer support pages for
individual processes. When the server terminates all of these processes
will also be cleaned up.