Update Puppeteer to v1.19.0 and Koa to 2.11.0
Since we now use alpine image for docker, we pin puppeteer to v1.19.0 as
it's recommended in Puppeteer docs.

By bumping Koa to v2 we had to update the code to use async/await.

A minor bug fix for the environment `SCREENIE_SCREENSHOT_DELAY` which now
uses regular `||` instead of the bitwise-or `|` syntax.