Update dependencies (#24)
Particularly Puppeteer has seen lots of updates since we last updated our
version of it. Mostly they've added support for Firefox, and lots of nice
new features we don't make use of yet, with no breaking changes for our
use. The main breaking change is a bump in minimum Node version, but we
are way overdo in that too anyway.

Together with Puppeteer, Koa and Winston have also received minor updates.